#135-create Activity module associated with no-arrows sessions
Completed. Still bugged.
